eunice-hong/src/test/kotlin/com/eunicehong/Day04Test.kt
package com.eunicehong

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

/**
 * Check if the solutions for [Day04] is correct.
 */
internal class Day04Test : DailyTest(4) {
    private val testDay04: Day04 = Day04()

    @Test
    override fun puzzleSolution1() {
        val expected = "2549"
        val actual = testDay04.solution1(input)
        assertEquals(expected, actual, "🎅 Solution for Puzzle 1 Day $dayString has not been implemented yet.")
        println("🌲 Day $dayString Puzzle 1 completed! $actual")
    }

    @Test
    override fun puzzleSolution2() {
        val expected = "Not implemented"
        val actual = testDay04.solution2(input)
        assertEquals(expected, actual, "🎅 Solution for Puzzle 2 Day $dayString has not been implemented yet.")
        println("🌲 Day $dayString Puzzle 2 completed! $actual")
    }
}
